+++ lib/api/src/Cardano/Wallet/Api/Http/Server/Handlers/TxCBOR.hs
@@ -18,6 +18,10 @@ import Prelude
 import Cardano.Binary
     ( DecoderError
     )
+import Cardano.Read.Ledger.Tx.CBOR
+    ( TxCBOR
+    , deserializeTx
+    )
 import Cardano.Wallet.Api.Http.Server.Error
     ( IsServerError (..)
     , apiError
@@ -44,10 +48,6 @@ import Cardano.Wallet.Read.Eras
     )
 import Cardano.Wallet.Read.Eras.EraFun
     ( applyEraFun
-    )
-import Cardano.Wallet.Read.Tx.CBOR
-    ( TxCBOR
-    , deserializeTx
     )
 import Cardano.Wallet.Read.Tx.Certificates
     ( getEraCertificates
