+++ lib/wallet/src/Cardano/Wallet.hs
@@ -279,6 +279,9 @@ import Cardano.Ledger.Api
 import Cardano.Mnemonic
     ( SomeMnemonic
     )
+import Cardano.Read.Ledger.Tx.CBOR
+    ( TxCBOR
+    )
 import Cardano.Slotting.Slot
     ( SlotNo (..)
     )
@@ -589,9 +592,6 @@ import Cardano.Wallet.Primitive.Types.UTxO
     )
 import Cardano.Wallet.Primitive.Types.UTxOStatistics
     ( UTxOStatistics
-    )
-import Cardano.Wallet.Read.Tx.CBOR
-    ( TxCBOR
     )
 import Cardano.Wallet.Shelley.Transaction
     ( _txRewardWithdrawalCost
